requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
h5py==3.9.0
gunicorn==21.2.0
